[DEV-757] moved masternode to infrastructure
