+++ src/subdomains/staking/infrastructre/masternode/api/controllers/masternode.controller.ts
@@ -3,12 +3,12 @@ import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { WalletRole } from 'src/shared/auth/wallet-role.enum';
-import { CreateMasternodeDto } from '../../application/dto/input/create-masternode.dto';
-import { PrepareResignMasternodeDto } from '../../application/dto/input/prepare-resign-masternode.dto';
-import { ResignMasternodeDto } from '../../application/dto/input/resign-masternode.dto';
+import { ResignMasternodeDto } from '../../application/dto/resign-masternode.dto';
 import { MasternodeService } from '../../application/services/masternode.service';
 import { Masternode } from '../../domain/entities/masternode.entity';
-import { MasternodeState } from '../../domain/enums';
+import { MasternodeState } from '../../../../domain/enums';
+import { CreateMasternodeDto } from '../../application/dto/create-masternode.dto';
+import { PrepareResignMasternodeDto } from '../../application/dto/prepare-resign-masternode.dto';
 
 @ApiTags('masternode')
 @Controller('masternode')
