+++ src/subdomains/staking/infrastructre/masternode/application/dto/create-masternode.dto.ts
