+++ src/subdomains/staking/infrastructre/masternode/application/dto/prepare-resign-masternode.dto.ts
