+++ src/subdomains/staking/infrastructre/masternode/application/dto/resign-masternode.dto.ts
