+++ src/subdomains/staking/infrastructre/masternode/application/repositories/masternode.repository.ts
